+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added collection support to `import-object` command.
+
 ### Fixed
 
 - Add newline to `get-payment-address` command output.